feat: move app types to shared package
