+++ apps/web/src/context/DataBlogProvider.tsx
@@ -1,7 +1,7 @@
 // src/context/DataBlogContext.tsx
 "use client";
 import React, { createContext, useContext, useState, useEffect, ReactNode, useMemo } from "react";
-import type { BlogData } from "@src/types/blog";
+import type { BlogData } from "@types/blog";
 
 // prettier-ignore
 const PUBLIC_DATA_URL = "https://amplify-d2jefuxcjjakai-ma-publiquestoragebucketac0-tjlluvtci6g6.s3.eu-west-3.amazonaws.com/publique-storage/data.json";
