--- apps/web/src/types/index.ts
@@ -1,7 +1,0 @@
-// @src/types/index.ts
-
-export * from "./blog";
-export * as CommentTypes from "./models/comment";
-export * as TodoTypes from "./models/todo";
-export * as UserNameTypes from "./models/userName";
-export * as UserProfileTypes from "./models/userProfile";
