+++ apps/web/src/utils/blogData/loadData.ts
@@ -1,6 +1,6 @@
 // src/utils/loadData.ts
 import { fetchBlogData } from "./fetchData";
-import type { Section, Post, Author } from "@src/types/blog";
+import type { Section, Post, Author } from "@types/blog";
 
 export async function loadData(): Promise<{
     sections: Section[];
