+++ packages/types/src/global.d.ts
