+++ packages/types/src/index.ts
@@ -1,1 +1,7 @@
-export {};
+// @types/index.ts
+
+export * from "./blog";
+export * as CommentTypes from "./models/comment";
+export * as TodoTypes from "./models/todo";
+export * as UserNameTypes from "./models/userName";
+export * as UserProfileTypes from "./models/userProfile";
